moved NameTag to app/components/UI
